Fix history

+++ src/reo7sp/cleverday/ui/activity/HistoryViewActivity.java
@@ -33,7 +33,7 @@ public class HistoryViewActivity extends Activity {
 		super.onCreate(savedInstanceState);
 
 		// parsing intent args
-		int id = getIntent().getIntExtra("id", -1);
+		long id = getIntent().getLongExtra("id", -1);
 		block = Core.getDataCenter().getBlock(id);
 		if (block == null) {
 			finish();
